requirements.txt
google-adk>=1.19.0
python-dotenv>=1.0.0
colorama>=0.4.6
pydantic>=2.0.0
